Prevent warning in standard deviation calculations
Due to roundoff errors in accumulation, the square of the sum of
values may be slightly larger that the sum of the squares of values
when the values are very small. As such, this changes rounds these
negative differences up to 0.0 in order to prevent taking the square
roots of negative numbers.